requirements.txt
numpy==1.16.5
scipy==1.0.1
matplotlib==2.2.2
tqdm==4.23.0
scikit-learn==0.20.3
scikit-image==0.14.3
pillow==6.0.0
SimpleITK==1.1.0
tensorflow-gpu==1.14.0
keras==2.3.1
h5py==2.7.1
opencv-python==4.1.1.26
